+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   generated. When the source version of input code is Java 9 or greater then the
   `javax.annotation.processing.Generated` annotation is added rather than the historic
   `javax.annotation.Generated` which can be difficult to support in Java 9 due to split modules.
+* **\[docs\]** Enable search on the website.
 
 ### [v0.67](https://github.com/react4j/react4j/tree/v0.67) (2018-04-20)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.66...v0.67)